api/requirements.txt
httpx==0.28.1
PyPDF2==3.0.1
python-docx==1.2.0
